+++ packages/forge-cli/src/cli/index.js
@@ -1,5 +1,5 @@
 exports.account = require('./account');
 exports.block = require('./block');
-exports.create = require('./create');
+// exports.create = require('./create');
 exports.node = require('./node');
 exports.tx = require('./tx');