+++ packages/forge-cli/src/cli/node/index.js
@@ -6,6 +6,7 @@ exports.init = require('./init');
 exports.ps = require('./ps');
 exports.reboot = require('./reboot');
 exports.restart = require('./restart');
+exports.simulate = require('./simulate');
 exports.start = require('./start');
 exports.state = require('./state');
 exports.stop = require('./stop');