+++ packages/forge-cli/src/cli/node/simulate/index.js
@@ -1,0 +1,18 @@
+// eslint-disable-next-line import/no-unresolved
+const { cli, action } = require('core/cli');
+const { execute, run } = require('./simulate');
+
+cli(
+  'simulate [action]',
+  'Start/stop simulator and generate some random data',
+  input => action(execute, run, input),
+  {
+    requirements: {
+      forgeRelease: true,
+      runningNode: true,
+      rpcClient: false,
+      wallet: false,
+    },
+    options: [],
+  }
+);